Add exception handling for "other" errors
For the actual HTTP requests, we can't rely on responses being
json. For example, maybe we get back an HTML body that explains a
gateway timeout.

In the previous version, this would make the actual error disappear
and just leave us with a JSONDecodeError.

If the JSON parsing fails, just fall back to using the raw response
body as the message for the Unclassified error.